Add sponsored event model
