+++ src/pycontw2016/settings/base.py
@@ -96,6 +96,7 @@ LOCAL_APPS = (
     'proposals',
     'users',
     'reviews',
+    'events',
     'sponsors',
 )
